Organize the code and get rid of unused bits
