--- src/ensime/formatter.properties
@@ -1,29 +1,0 @@
-# formatter preferences
-
-# Indentation & Alignment
-indentSpaces=2
-indentWithTabs=false
-alignParameters=true
-doubleIndentClassDeclaration=true
-alignSingleLineCaseStatements=true
-alignSingleLineCaseStatements.maxArrowIndent=40
-indentPackageBlocks=true
-indentLocalDefs=false
-
-# Spaces
-spaceBeforeColon=false
-compactStringConcatenation=false
-spaceInsideBrackets=false
-spaceInsideParentheses=false
-preserveSpaceBeforeArguments=false
-spacesWithinPatternBinders=true
-
-# Scaladoc
-multilineScaladocCommentsStartOnFirstLine=false
-placeScaladocAsterisksBeneathSecondAsterisk=false
-
-# Miscellaneous
-formatXml=true
-rewriteArrowSymbols=false
-preserveDanglingCloseParenthesis=false
-compactControlReadability=false
